+++ app/globals.css
@@ -42,7 +42,7 @@ body, main {
   width: 100%;
   height: 100%;
   top: 0;
-  opacity: 0.4;
+  opacity: 0.4 ;
   filter: invert(1);
 }
